Create infographics web site

Create infographics web site

create_info_site(
  path_root = ".",
  dir_svg = "svg",
  elements_csv = "svg_elements.csv",
  indicators_csv = "plot_indicators.csv",
  site_title = "MBON Interactive Infographics",
  plot_function = "plot_dygraph_timeseries",
  dir_rmd = "rmd",
  dir_web = "docs",
  svg_paths = list.files(file.path(path_root, dir_svg), ".*\\.svg$", full.names = T),
  svg_names = tools::file_path_sans_ext(basename(svg_paths)),
  site_yml_brew = system.file("site_template/_site.yml.brew", package = "infographiq"),
  index_md_brew = system.file("site_template/index.md.brew", package = "infographiq"),
  readme_md_brew = system.file("site_template/README.md.brew", package = "infographiq"),
  header = system.file("site_template/_header.html", package = "infographiq"),
  footer = system.file("site_template/_footer.html", package = "infographiq"),
  styles_css = system.file("site_template/styles.css", package = "infographiq"),
  render_modals = T,
  preview_site = T
)

Arguments

path_root
root path for repository, from which other directories are relatively positioned.

dir_svg
directory within path_root containing SVG files.

elements_csv
csv within path_root containing SVG identifiers and having these columns: svg (filename of svg, eg pelagic.svg), svg_id (SVG id of element), label (shows on hover and in modal title), status_text (shows on hover, optional), status_color (hexadecimal color to fill shape, optional).

indicators_csv
csv table containing pointers going from svg_id to csv_url with data for plot of indicator for use with plot_dygraph_timeseries() function.

site_title
title of site to show up in navigation menu.

plot_function
plot function to use (defaults to plot_dygraph_timeseries)

dir_rmd
directory within path_root to output Rmarkdown files for staging website.

dir_web
directory within path_root to output html files for final website. Defaults to 'docs' in order to comply with Github Pages repository configuration.

svg_paths
paths to specific svg files. Defaults to all svg files in dir_svg.

svg_names
names for each svg scene that becomes a menu item in site. Defaults to name of svg with file extension (.svg) stripped.

site_yml_brew
brew template for site menu to render the Rmarkdown website (see http://rmarkdown.rstudio.com/rmarkdown_websites.html).

index_md_brew
brew template for home page of site

readme_md_brew
brew template for README of site, accessible via upper right "?" icon in menu

header
header for site

footer
footer for site

styles_css
cascading style sheet (css) for site

render_modals
True/False on whether to render the modal time series plots for site. Can be time consuming and gets cached in the dir_rmd folder, so set this to False if working on other parts of already generated site.

preview_site
True/False on whether to launch website in RStudio. Because of cross-browser security restrictions some SVG functionality of the site does not work just opening the html files locally from the filesystem, so requires a hosted web server either remotely (eg Github.io) or locally (ie servr::httd()).

Value

nothing returned from the function if no error, except the created site is generated.

Examples

library(infographiq)

#> 
#> Attaching package: ‘infographiq’

#> The following objects are masked from ‘package:infographiqR’:
#> 
#>     create_info_site, get_plotting_function_brew, info_svg, modal_html,
#>     plot_dygraph_timeseries, plot_intertidal_nms

#create_info_site()